+++ addons/storyshots/storyshots-core/src/api/getIntegrityOptions.js
