+++ addons/storyshots/storyshots-core/src/frameworks/angular/app.component.ts
