+++ addons/storyshots/storyshots-core/src/frameworks/angular/app.token.ts
