+++ addons/storyshots/storyshots-core/src/frameworks/angular/loader.js
