+++ addons/storyshots/storyshots-core/src/frameworks/hasDependency.js
