+++ addons/storyshots/storyshots-core/src/frameworks/html/loader.js
