+++ addons/storyshots/storyshots-core/src/frameworks/react/loader.js
