+++ addons/storyshots/storyshots-core/src/frameworks/react/renderShallowTree.js
