+++ addons/storyshots/storyshots-core/src/frameworks/vue/loader.js
